Updated enhancement middleware and middleare module structure
