--- src/middleware/_module-exports.ts
@@ -1,3 +1,0 @@
-export type * from "@/middleware/types.js";
-export * from "@/middleware/enhance-factory.js";
-export * from "@/middleware/use-factory.js";
