+++ src/middleware/implementations/enhance-factory/_module.ts
@@ -1,0 +1,1 @@
+export * from "@/middleware/implementations/enhance-factory/enhance-factory.js";
