+++ src/middleware/implementations/use-factory/_module.ts
@@ -1,0 +1,1 @@
+export * from "@/middleware/implementations/use-factory/use-factory.js";
